requirements.txt
gradio
pillow
torch
torchvision
numpy
ddgs
playsound
requests
opencv-python
